+++ pkg/search/ann_profile.go
@@ -82,7 +82,7 @@ func ResolveCompressedANNProfile(vectorCount, dimensions int, vectorStoreReady b
 	profile.PQBits = clampInt(envutil.GetInt("NORNICDB_VECTOR_PQ_BITS", 8), 4, 8)
 	defaultNProbe := maxInt(16, (profile.IVFLists+7)/8)
 	profile.NProbe = clampInt(envutil.GetInt("NORNICDB_VECTOR_IVFPQ_NPROBE", defaultNProbe), 1, profile.IVFLists)
-	profile.RerankTopK = clampInt(envutil.GetInt("NORNICDB_VECTOR_IVFPQ_RERANK_TOPK", 200), 10, 20000)
+	profile.RerankTopK = clampInt(envutil.GetInt("NORNICDB_VECTOR_IVFPQ_RERANK_TOPK", 2000), 10, 20000)
 	profile.TrainingSampleMax = clampInt(envutil.GetInt("NORNICDB_VECTOR_IVFPQ_TRAINING_SAMPLE_MAX", 200000), 1000, 5000000)
 	profile.OverflowMax = clampInt(envutil.GetInt("NORNICDB_VECTOR_IVFPQ_OVERFLOW_MAX", 512), 0, 100000)